+++ src/statefunctions/gamestates/ron.cc
@@ -3,7 +3,7 @@
 #include <memory>
 #include <vector>
 
-#include "analysis/scoring.h"
+#include "scoring/scoring.h"
 #include "statefunctions/router.h"
 #include "statefunctions/stateutilities.h"
 #include "types/event.h"
