+++ tests/yakus/allterminalsandhonors.test.cc
@@ -5,7 +5,7 @@
 #include <vector>
 
 #include "analysis/analysis.h"
-#include "analysis/yakus.h"
+#include "scoring/yakus.h"
 
 #include "types/gamestate.h"
 #include "types/hand.h"
