+++ tests/yakus/puredoublechi.test.cc
@@ -4,7 +4,7 @@
 #include <string>
 
 #include "analysis/analysis.h"
-#include "analysis/yakus.h"
+#include "scoring/yakus.h"
 
 #include "types/gamestate.h"
 #include "types/hand.h"
